Design Pattern
1. Template Method Pattern
2. TransportState.java, NetworkState.java
3. Now there is only TransportState but aftertime, doing maintenance, there could be more State. So we make NetworkState abstract class to make managing duplicate code easier and improve software maintainability.
